// src/schemas/balancer-api.ts
export interface paths {
// ...
	"/gateways/{uuid}/config": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * Return YAML config for Traefik Gateway
		 * @description This endpoint is called from CloudGateway Automation only. So we are not going to provide ability to Account Admins to see a Gateway Config
		 *
		 */
		get: {
			parameters: {
				query?: never;
				header?: never;
				path: {
					/** @description Used for simplified authorization for some internal requests from CloudGateway to config.
					 *      */
					uuid: components["parameters"]["CloudGatewayId"];
				};
				cookie?: never;
			};
			requestBody?: never;
			responses: {
				/** @description Operation successful */
				200: {
					headers: {
						[name: string]: unknown;
					};
					content: {
						"text/yaml": components["schemas"]["GatewayConfig"];
					};
				};
				404: components["responses"]["not_found"];
			};
		};
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
	"/certificates/{id}": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * Show certificate record
		 * @description This endpoint is used to explore the details of a certificate.
		 *     Role-based access:
		 *       - `operator` role can see all certificates in location
		 *       - `admin`/`owner` role can see all certificates in account
		 *       - `member` role can see only own certificates
		 *
		 */
		get: operations["certificate_show"];
		/**
		 * update cert
		 * @description The endpoint is called for updating a certificate. The server automatically checks the validity of the updated certificate. After the certificate has been successfully updated via the API, CloudGateway Automation automatically downloads and applies it to a route defined in the hostnames field.
		 *     Role-based access:
		 *       - `operator` role can manage all certificates in location
		 *       - `admin`/`owner` role can manage all certificates in account
		 *       - `member` role can manage only own certificates
		 *
		 */
		put: operations["certificate_update"];
		post?: never;
		/**
		 * Delete certificate
		 * @description The endpoint is called to delete a certificate from the system.
		 *     Role-based access:
		 *       - `operator` role can manage all certificates in location
		 *       - `admin`/`owner` role can manage all certificates in account
		 *       - `member` role can manage only own certificates
		 *
		 */
		delete: operations["certificate_delete"];
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
}
export type webhooks = Record<string, never>;
export interface components {
// ...
	responses: {
		/** @description This API gateway error means selected operation is not supported */
		apigw_not_allowed: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": string;
			};
		};
		/** @description Server error. */
		server_error: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Request is not authorized, usually because of: bad credentials, expired JWT token, account role is not assigned to authenticated user.
		 *      */
		unauthorized: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Request authorized, but current user doesn't have access rights to perform action.
		 *      */
		forbidden: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Requested object is not found
		 *      */
		not_found: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description Unprocessable entity.
		 *      */
		unprocessable_entity: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["error"];
			};
		};
		/** @description A list of ALB Gateways attached to account's infrastructure */
		alb_list: {
			headers: {
				[name: string]: unknown;
			};
			content: {
				"application/json": components["schemas"]["GatewayCollection"];
			};
		};
	};
	parameters: {
		/** @description Name of the account (5-character string) */
		Account: string;
		/** @description Role name */
		Role: string;
		/** @description Used for simplified authorization for some internal requests from CloudGateway to config.
		 *      */
		CloudGatewayId: string;
	};
	requestBodies: never;
	headers: never;
	pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
	alb_list: {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		requestBody?: never;
		responses: {
			200: components["responses"]["alb_list"];
			401: components["responses"]["unauthorized"];
			500: components["responses"]["server_error"];
		};
	};
	alb_attach: {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		requestBody: {
			content: {
				"application/json": {
					/** @example alb-default */
					name?: string;
					/** @example a843bf2a-7353-563e-a220-f59949849dba */
					cloudgw_id?: string;
					/** @example acc */
					account?: string;
				};
			};
		};
		responses: {
			200: components["responses"]["alb_list"];
			401: components["responses"]["unauthorized"];
			405: components["responses"]["apigw_not_allowed"];
			422: components["responses"]["unprocessable_entity"];
		};
	};
// ...
}